+++ tests/phpunit/Unit/ImageModalTest.php
@@ -386,11 +386,18 @@ class ImageModalTest extends PHPUnit_Framework_TestCase {
 	public function canParseDataProvider() {
 		$globalConfig = MediaWikiServices::getInstance()->getMainConfig();
 		$scriptPath = $globalConfig->get( 'ScriptPath' );
+		# @todo remove this, when dropping support for 1.31
+		$injectedAsyncLoading = (
+			version_compare( $globalConfig->get( 'Version' ), '1.32', 'gt' )
+				? 'decoding="async" '
+				: ''
+		);
 		/*
 		 * notes on adding tests:
 		 * - when using manual thumbnail, inject $scriptPath: <img alt="" src="' . $scriptPath . '/images/a/aa/Shuttle.png" width="68" height="18" ...
 		 * - always supply an align value, otherwise testing will fail with an exception due to bad class design (blame @oetterer)
-		 * - switch values (booleans) are true when present and false. see "frameless" on test "manual width, frameless"
+		 * - switch values (booleans) are true when not present and false. see "frameless" on test
+		 * "manual width, frameless"
 		 */
 		return [
 			'no params' => [
@@ -447,7 +454,7 @@ class ImageModalTest extends PHPUnit_Framework_TestCase {
 					'framed'      => false,
 				],
 				[],
-				'<div class="thumb tnone"><span class="modal-trigger" data-toggle="modal" data-target="#bsc_image_modal_test"><div class="thumbinner" style="width:70px;"><img alt="" src="' . $scriptPath . '/images/a/aa/Shuttle.png" decoding="async" width="68" height="18" class="thumbimage" />  <div class="thumbcaption"></div></div></span></div>',
+				'<div class="thumb tnone"><span class="modal-trigger" data-toggle="modal" data-target="#bsc_image_modal_test"><div class="thumbinner" style="width:70px;"><img alt="" src="' . $scriptPath . '/images/a/aa/Shuttle.png" ' . $injectedAsyncLoading . 'width="68" height="18" class="thumbimage" />  <div class="thumbcaption"></div></div></span></div>',
 				'<div class="modal fade" role="dialog" id="bsc_image_modal_test" aria-hidden="true"><div class="modal-dialog modal-lg"><div class="modal-content"><div class="modal-header"><button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button></div><div class="modal-body"><img src=TEST_OUTPUT class="img-fluid"></div>'
 				. '<div class="modal-footer"><a class="btn btn-primary" role="button" href="/File:Serenity.png">Visit Source</a><button type="button" class="btn btn-default" data-dismiss="modal" aria-label="Close">Close</button></div></div></div></div>' . "\n",
 			],
@@ -479,7 +486,7 @@ class ImageModalTest extends PHPUnit_Framework_TestCase {
 					'manualthumb' => 'Shuttle.png',
 				],
 				[],
-				'<div class="thumb tleft"><span class="modal-trigger" data-toggle="modal" data-target="#bsc_image_modal_test"><div class="thumbinner" style="width:70px;"><img alt="" src="' . $scriptPath . '/images/a/aa/Shuttle.png" decoding="async" width="68" height="18" class="thumbimage" />  <div class="thumbcaption"><div class="magnify"><a class="internal" title="Enlarge"></a></div></div></div></span></div>',
+				'<div class="thumb tleft"><span class="modal-trigger" data-toggle="modal" data-target="#bsc_image_modal_test"><div class="thumbinner" style="width:70px;"><img alt="" src="' . $scriptPath . '/images/a/aa/Shuttle.png" ' . $injectedAsyncLoading . 'width="68" height="18" class="thumbimage" />  <div class="thumbcaption"><div class="magnify"><a class="internal" title="Enlarge"></a></div></div></div></span></div>',
 				'<div class="modal fade" role="dialog" id="bsc_image_modal_test" aria-hidden="true"><div class="modal-dialog modal-lg"><div class="modal-content"><div class="modal-header"><button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>'
 				. '<div class="modal-body"><img src=TEST_OUTPUT class="img-fluid"></div><div class="modal-footer"><a class="btn btn-primary" role="button" href="/File:Serenity.png">Visit Source</a><button type="button" class="btn btn-default" data-dismiss="modal" aria-label="Close">Close</button></div></div></div></div>' . "\n",
 			],