+++ tests/phpunit/Unit/ImageModalTriggerTest.php
@@ -124,6 +124,13 @@ class ImageModalTriggerTest extends PHPUnit_Framework_TestCase {
 	public function canGenerateProvider() {
 		$globalConfig = MediaWikiServices::getInstance()->getMainConfig();
 		$scriptPath = $globalConfig->get( 'ScriptPath' );
+		# @todo remove this, when dropping support for 1.31
+		$injectedAsyncLoading = (
+		version_compare( $globalConfig->get( 'Version' ), '1.32', 'gt' )
+			? 'decoding="async" '
+			: ''
+		);
+
 		/*
 		 * 1. Parameter: Sanitized frame parameter:
 		 * All of these must be present and true or false:
@@ -241,7 +248,7 @@ class ImageModalTriggerTest extends PHPUnit_Framework_TestCase {
 				[
 					'page' => false,
 				],
-				'<div class="thumb tnone"><span class="modal-trigger" data-toggle="modal" data-target="#id"><div class="thumbinner" style="width:70px;"><img alt="" src="' . $scriptPath . '/images/a/aa/Shuttle.png" decoding="async" width="68" height="18" class="thumbimage" />  <div class="thumbcaption"></div></div></span></div>',
+				'<div class="thumb tnone"><span class="modal-trigger" data-toggle="modal" data-target="#id"><div class="thumbinner" style="width:70px;"><img alt="" src="' . $scriptPath . '/images/a/aa/Shuttle.png" ' . $injectedAsyncLoading. 'width="68" height="18" class="thumbimage" />  <div class="thumbcaption"></div></div></span></div>',
 			],
 			'framed'                         => [
 				[
@@ -298,7 +305,7 @@ class ImageModalTriggerTest extends PHPUnit_Framework_TestCase {
 				[
 					'page' => false,
 				],
-				'<div class="thumb tleft"><span class="modal-trigger" data-toggle="modal" data-target="#id"><div class="thumbinner" style="width:70px;"><img alt="" src="' . $scriptPath . '/images/a/aa/Shuttle.png" decoding="async" width="68" height="18" class="thumbimage" />  <div class="thumbcaption"><div class="magnify"><a class="internal" title="Enlarge"></a></div></div></div></span></div>',
+				'<div class="thumb tleft"><span class="modal-trigger" data-toggle="modal" data-target="#id"><div class="thumbinner" style="width:70px;"><img alt="" src="' . $scriptPath . '/images/a/aa/Shuttle.png" ' . $injectedAsyncLoading . 'width="68" height="18" class="thumbimage" />  <div class="thumbcaption"><div class="magnify"><a class="internal" title="Enlarge"></a></div></div></div></span></div>',
 			],
 		];
 	}